rs/rapido-vm: split out rapido::host_kernel_vers into lib
rapido-cut needs the same functionality, so split it out from rapido-vm.
