+++ src/lib.rs
@@ -1,0 +1,35 @@
+// SPDX-License-Identifier: (GPL-2.0 OR GPL-3.0)
+// Copyright (C) 2025 SUSE LLC
+use std::io;
+use std::fs;
+
+// parse /proc/version string, e.g. Linux version 6.17.0-2-default ...
+fn host_kernel_vers_parse(kvers: &[u8]) -> io::Result<String> {
+    match str::from_utf8(kvers) {
+        Err(_) => Err(io::Error::from(io::ErrorKind::InvalidData)),
+        Ok(s) => match s.strip_prefix("Linux version ") {
+            None => Err(io::Error::from(io::ErrorKind::InvalidData)),
+            Some(rel) => match rel.split_once([' ']) {
+                Some((rel, _)) => Ok(rel.to_string()),
+                None => Err(io::Error::from(io::ErrorKind::InvalidData)),
+            },
+        },
+    }
+}
+
+// return the host kernel version based on /proc/version contents
+pub fn host_kernel_vers() -> io::Result<String> {
+    let kvers = fs::read("/proc/version")?;
+    host_kernel_vers_parse(&kvers)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_host_kernel_vers_parse() {
+        let line = b"Linux version 6.17.0-2-default (geeko@buildhost) (gcc (SUSE Linux) 15.2.0, GNU ld (GNU Binutils; openSUSE Tumbleweed) 2.43.1.20241209-10) #1 SMP PREEMPT_DYNAMIC Thu Oct  2 08:12:40 UTC 2025 (190326b)";
+        assert_eq!(host_kernel_vers_parse(line).unwrap(), "6.17.0-2-default");
+    }
+}
